% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abrev_mois.R
\encoding{UTF-8}
\name{abrev_mois}
\alias{abrev_mois}
\title{Utils}
\usage{
abrev_mois(mois)
}
\arguments{
\item{mois}{Nombre entier entre 1 et 12 en référence aux 12 mois de l'année.}
}
\value{
\code{string}
}
\description{
Utils
}
\keyword{internal}
