% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DER_inesss_update.R
\encoding{UTF-8}
\name{DER_inesss_update.addins}
\alias{DER_inesss_update.addins}
\title{Install package inesss}
\usage{
DER_inesss_update.addins()
}
\arguments{
\item{direction}{Nom de la direction.}
}
\description{
Installation ou mise à jour de la librairie selon la direction indiquée.
}
\keyword{internal}
