% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DER_inesss_update.R
\name{DER_inesss_update}
\alias{DER_inesss_update}
\title{Install package inesss}
\usage{
DER_inesss_update()
}
\arguments{
\item{direction}{Nom de la direction.}
}
\description{
Installation ou mise à jour de la librairie selon la direction indiquée.
}
