% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/from_bd.vue.R
\encoding{UTF-8}
\name{from_bd.vue}
\alias{from_bd.vue}
\title{Utils}
\usage{
from_bd.vue(bd = "PROD", vue)
}
\arguments{
\item{bd}{Nom de la base de données. Souvent \code{PROD}.}

\item{vue}{Nom de la vue.}
}
\value{
"from \code{bd}.\code{vue}"
}
\description{
FROM clause de SQL
}
\keyword{internal}
