% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fusion_periodes.R
\name{fusion_periodes}
\alias{fusion_periodes}
\title{Arrangement périodes}
\usage{
fusion_periodes(dt, debut, fin, cols, njours = 1L)
}
\arguments{
\item{dt}{Data à arranger.}

\item{debut}{Nom de la colonne indiquant le début de la période.}

\item{fin}{Nom de la colonne indiquant la fin de la période.}

\item{cols}{Nom de la ou des colonnes indiquant en quoi consiste la période (codes d'identification d'individus ou de produits).}

\item{njours}{Nombre de jours entre le début et la fin précédente (debut{i} - fin{i-1}). Voir examples. Par défaut 1.}
}
\value{
\code{data.table}
}
\description{
Fusionne les périodes qui se chevauchent dans le temps.
}
\details{
La variable \code{cols} ne peut contenir les valeurs "debut" ou "fin".
}
\examples{
require(data.table)
ex1 <- data.frame(
  ID = c(1, 1, 1, 2, 2),
  DATE_DEBUT = as.Date(c("2020-03-01", "2020-04-01", "2020-05-02",
                         "2020-06-15", "2020-06-22")),
  DATE_FIN = as.Date(c("2020-03-31", "2020-04-30", "2020-05-31",
                       "2020-06-25", "2020-06-30"))
)
ex1_1 <- fusion_periodes(ex1, debut = "DATE_DEBUT", fin = "DATE_FIN", cols = "ID", njours = 1L)
}
