% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArgCheck.R
\encoding{UTF-8}
\name{newArgCheck}
\alias{newArgCheck}
\title{Argcheck}
\usage{
newArgCheck()
}
\value{
environment
}
\description{
Create an environment where to put errors, messages or warnings.
}
\keyword{internal}
