% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmNA.R
\encoding{UTF-8}
\name{rmNA}
\alias{rmNA}
\title{Astuce}
\usage{
rmNA(x)
}
\arguments{
\item{x}{Vecteur.}
}
\description{
Supprime les \code{NA}s du vecteur. Renvoie \code{NULL} si aucune valeur.
}
\examples{
rmNA(c(4, 6, 8, NA, 78, 4, NaN))
}
